requirements.txt
setuptools~=80.9.0
python-telegram-bot~=20.7
gTTS~=2.5.1
platformdirs~=4.5.1